requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
mysql-connector-python==8.1.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
textblob==0.17.1
nltk==3.8.1
joblib==1.3.2
werkzeug==2.3.7
bcrypt==4.0.1
email-validator==2.0.0
reportlab==4.0.4